move convertBytesToHumanReadable to file util
