+++ src/services/upload/uploader.ts
@@ -13,7 +13,7 @@ import { FILE_TYPE } from 'constants/file';
 import { UPLOAD_RESULT, MAX_FILE_SIZE_SUPPORTED } from 'constants/upload';
 import { FileWithCollection, BackupedFile, UploadFile } from 'types/upload';
 import { addLogLine } from 'utils/logging';
-import { convertBytesToHumanReadable } from 'utils/billing';
+import { convertBytesToHumanReadable } from 'utils/file';
 import { sleep } from 'utils/common';
 import { addToCollection } from 'services/collectionService';
